refactor(tui): extract system theme generation
