+++ packages/tui/test/theme.test.ts
@@ -2,8 +2,9 @@ import { expect, test } from "bun:test"
 import { mkdir, writeFile } from "node:fs/promises"
 import path from "node:path"
 import type { TerminalColors } from "@opentui/core"
-import { DEFAULT_THEMES, addTheme, allThemes, hasTheme, resolveTheme, terminalMode } from "../src/theme"
+import { DEFAULT_THEMES, addTheme, allThemes, hasTheme, resolveTheme } from "../src/theme"
 import { discoverThemes } from "../src/context/theme"
+import { terminalMode } from "../src/theme/system"
 import { tmpdir } from "./fixture/fixture"
 
 test("addTheme writes into module theme store", () => {
